Add wire.ts
